Fix datetime bug
